"Add cart component with modal and move meals component up in the hierarchy"
This diff adds a new Cart component to the application, which is displayed in the App component. The Cart component includes a Modal component, which is implemented in a separate Modal.js file. The Meals component is also moved up in the hierarchy and is now displayed below the Header component. Additionally, the Cart component is added to the overlays element in the index.html file using a portal.