"Added onClick prop to HeaderCartButton"
Explanation:
The code diff shows that the `HeaderCartButton` component's `button` element now has an `onClick` prop added to it. However, the value of the `onClick` prop is an empty object literal `{{}}`, which means that it is not yet functional. Nonetheless, the change is still significant because it lays the groundwork for adding functionality to the button in the future.