"Add cart reduction logic for ADD action type"
This commit adds new code to the cartReducer function to handle an 'ADD' action type. It creates a new array of items by concatenating the existing items with the new item to be added. It also calculates the new total amount by adding the price and amount of the new item to the existing total amount. The updated items and total amount are then returned as the new state.