"Change functional component to class component for Input"
Explanation:
The code diff shows that the Input component has been changed from a functional component to a class component using React.createRef. However, the component still uses functional syntax for rendering the input element. This change might be a step towards using class component features like state or lifecycle methods, but as it is, it only introduces a different way of defining the component.