"Added state for amount validation and conditional rendering of error message in MealItemForm"
Explanation:
The code diff shows the addition of a state variable `amountIsValid` and its updater function `setAmountIsValid` in the `MealItemForm` component. Also, a conditional rendering of an error message is added based on the value of `amountIsValid`. This suggests that the component now has the ability to validate the input amount and display an error message if the amount is not within the valid range (1-5).