"Add CartItem component to Cart and pass necessary props"
This commit adds the `CartItem` component to the `Cart` component and passes the necessary props to it. The `cartItem` variable in the `Cart` component is updated to map over `cartCtx.items` and render a `CartItem` component for each item, passing the `name`, `price`, `amount`, `onRemove`, and `onAdd` props to it. The `onRemove` and `onAdd` props are currently empty functions and will need to be implemented in a future commit.