"Update cart reducer to merge items with same id and add new items to cart"
This diff updates the cart reducer to handle adding items to the cart more accurately. If an item with the same id already exists in the cart, the reducer will now merge the existing item with the new one by adding their quantities together. If the item does not already exist in the cart, it will be added as a new item as before. This change is implemented by introducing a new variable `existingCartItem` to check if an item with the same id already exists in the cart, and then updating the `updatedItems` variable accordingly.